Split out test into partial classes
